cleanup and refactoring based on review
